Support for writing explicit equivalences.
Introduces support for adapters to write explicit equivalences for
Content. References in a Content's 'equivalentTo' set are written to its
lookup entry in a field called 'explicitEquivalents' (analagous to
'directEquivalents' for *generated* equivalences.) Explicit equivalences
form part of the transitive equivalence graph just as generated
equivalences.

The 'equivalentTo' set of Identified is now of type LookupRef rather
than String as each equivalent requires a Publisher for writing
equivalecs. LookupRef has been moved to atlas-model to support this
change.