Ensure schedule item is at head of merged equivalents.
When equivalence merging is active when resolving a schedule the item in
the schedule must be selected as the Content into which others are
merged so the correct Broadcast is present in the merged result and the
Item is not stripped from the schedule.

This ensures the schedule item is at the head of the iterable of
Content for merging which, as the sort in the equivalence merging is
stable, means it'll be selected as precedent. The Set of equvialents is
rebuilt as an ImmutableSet with the schedule Item added first.