Fix running of MongoChannelStoreWriteTest with other test cases
This test was failing because the test mongo and channel stores
were static final variables and initialised at class loading time.
Therefore the initialisation of database performed in
MongoSequentialIdGenerator was being lost by other classes
subsequently getting a test mongo, which drops the db, ahead of
the tests in MongoChannelStoreWriteTest being run.