Mongo collection for equivalence writing should read only primary
Configure the `MongoLookupEntryStore` passed to the
`EquivalenceWritingContentWriter` with a mongo collection which will
only read from the *primary* replica.

This avoids a race in the front-end when POSTing Content where a
LookupEntry for the new content is written to the primary but when
recording its equivalence a read from the secondary for that LookupEntry
returns nothing as the replica is (even slightly) out of sync.