Filter out unpublished refs in getEquivSetByFollowingLinks
The equivs returned by this method would include unpublished content
along with potentially published content that is equived to the
unpublished content. As a result you could see published content
returned that is not reachable except through a link on unpublished
content.
This change filters out unpublished content before querying for
its direct/explicit equivalences, and filters out the LookupRefs it
had found when returning.