Try to trigger an equiv update when encountering unpublished content
When using the respect_api_key_for_equiv_list annotation, if we
come across unpublished content in the equiv set, write a message
on the content changes queue to try and have equivalence update
on the unpublished content, to try and remove it from the set
for future benefit.
This was added at Andreas's request to try and clean up equiv sets
where we encounter them.