Fixed a test in DefaultEquivalentContentResolverTest
Due to changes in DefaultEquivalentContentResolver the
content is now queried for a ref before the ref is added
to a map as opposed to immediately adding the ref to a set.
As a result the test is updated to store a lookup entry
of a piece of content that now needs to be resolved.