Changed respect-api-key annotation to always include the queried content
If the main content queried is unpublished we would return an empty
set of LookupRefs in getEquivSetByFollowingLinks. Since existing
behaviour for this annotation, as well as without this annotation
would still return the content (and possibly its equivs) even if
the content is unpublished, we return at least the LookupRef for
the queried content without its equivalences if the content is
unpublished.