Transitive lookup writer transactions (#350) (ENG-744)
Prevent race conditions when updating equiv sets in the TransitiveLookupWriter by using Mongo transactions to update all lookup entries as a single operation. The update will fail and retry if any entry that will be updated in the transaction is also modified by another process in the meantime.

As of Mongo 4.0 there is a size restriction on how large a transaction can be; if this occurs the update will be retried without using a transaction.

New Mongo driver classes supporting transactions have been added to the MongoLookupEntryStore to support this.