Return any state change in transitive lookup writer
The existing behaviour where the subset's update
was completely ignored was not beneficial and
flawed. It does not look like the return of this
method is used for anything other than to
detect if some change has happened (not necessarily
a complete change) and so this behaviour should
be fine to modify.

Updated and cleaned up relevant unit tests.